requirements.txt
fastembed-gpu
pillow
numpy
tqdm